+++ Panels/Header.js
@@ -346,11 +346,13 @@ const HeaderBase = kind({
 	},
 
 	computed: {
-		className: ({centered, children, noSubtitle, type, shadowed, styler, subtitle}) => styler.append(
+		className: ({backButtonAvailable, centered, children, noBackButton, noCloseButton, noSubtitle, type, shadowed, slotAfter, slotBefore, styler, subtitle}) => styler.append(
 			{
 				centered,
 				noSubtitle,
 				shadowed,
+				slotAfter: slotAfter || !noCloseButton,
+				slotBefore: slotBefore || (backButtonAvailable && !noBackButton),
 				// This likely doesn't need to be as verbose as it is, with the first 2 conditionals
 				withChildren: hasChildren(children),
 				withSubtitle: subtitle
@@ -458,7 +460,6 @@ const HeaderBase = kind({
 				iconFlip="auto"
 				onClick={onBack}
 				shadowed={shadowed}
-				size="small"
 			/>
 		) : null);
 
@@ -471,7 +472,6 @@ const HeaderBase = kind({
 				icon="closex"
 				onClick={onClose}
 				shadowed={shadowed}
-				size="small"
 			/>
 		) : null);