+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ### Changed
 
-- `limestone/Chip` styling to match the latest GUI
+- `limestone/Chips.Chip` styling to match the latest GUI
+- `limestone/TabLayout` styling to match the latest GUI
 
 ## [1.0.0-rc.2] - 2025-06-04