+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Changed
 
 - `limestone/Input` styling to match the latest GUI
+- `limestone/TabLayout` styling to match the latest GUI
 
 ## [1.0.0-rc.2] - 2025-06-04