+++ CHANGELOG.md
@@ -4,8 +4,13 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `limestone/Chips`, a new container component that groups and manages multiple `limestone/Chips.Chip` for improved layout and interaction
+
 ### Changed
 
+- `limestone/Chips.Chip` styling to match the latest GUI
 - `limestone/TabLayout` styling to match the latest GUI
 
 ## [1.0.0-rc.2] - 2025-06-04