+++ CHANGELOG.md
@@ -4,13 +4,48 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `limestone/Chips`, a new container component that groups and manages multiple `limestone/Chips.Chip` for improved layout and interaction
+
+### Changed
+
+- `limestone/Chips.Chip` styling to match the latest GUI
+- `limestone/ImageItem` styling to match the latest GUI
+- `limestone/TabLayout` styling to match the latest GUI
+
+## [1.0.0-rc.2] - 2025-06-04
+
+### Changed
+
+- `limestone/ContextualMenuDecorator` styling to match the latest GUI
+- `limestone/FixedPopupPanels` styling to match the latest GUI
+
+### Fixed
+
+- `limestone/TabLayout` to focus to the last focused tab when focus enters to the tab group
+
+## [1.0.0-rc.1] - 2025-05-29
+
+### Added
+
+- `limestone/Icon` supported icon list, adding new icons
+- `limestone/TabLayout` prop `primaryIndex` to focus primary tab when pressing back key from other tabs or initial rendering
+
 ### Changed
 
 - `limestone/BodyText` styling to match the latest GUI
+- `limestone/DatePicker` styling to match the latest GUI
 - `limestone/DayPicker` styling to match the latest GUI
-- `limestone/ImageItem` styling to match the latest GUI
+- `limestone/Heading` prop `size` default value to `tiny` and styling to match the latest GUI
+- `limestone/PageViews` styling to match the latest GUI
+- `limestone/Panels` styling to match the latest GUI
+- `limestone/Panels.Header` styling to match the latest GUI
+- `limestone/PopupTabLayout` styling to match the latest GUI
 - `limestone/ProgressButton` styling to match the latest GUI
 - `limestone/Steps` styling to match the latest GUI
+- `limestone/Switch` styling to match the latest GUI
+- `limestone/SwitchItem` styling to match the latest GUI
 - `limestone/WizardPanels` styling to match the latest GUI
 
 ### Fixed